Various changes to reestablish 4.0 backwards compat
* Revert NpgsqlStatement.Rows from ulong back to uint (#2150)
* Revert NpgsqlBinaryImporterComplete to return void (#2112)
* Moved NpgsqlTimeout from namespace Npgsql.Util back to Npgsql
* Various other minor missing cancellation token overloads

Fixes #2626
Fixes #2655
Fixes #2642
