+++ src/Npgsql/NpgsqlStatement.cs
@@ -33,7 +33,7 @@ namespace Npgsql
         /// See the command tag in the CommandComplete message,
         /// http://www.postgresql.org/docs/current/static/protocol-message-formats.html
         /// </remarks>
-        public ulong Rows { get; internal set; }
+        public uint Rows { get; internal set; }
 
         /// <summary>
         /// For an INSERT, the object ID of the inserted row if <see cref="Rows"/> is 1 and
@@ -101,7 +101,8 @@ namespace Npgsql
         internal void ApplyCommandComplete(CommandCompleteMessage msg)
         {
             StatementType = msg.StatementType;
-            Rows = msg.Rows;
+            // Downcast to uint for backwards compat with 4.0
+            Rows = (uint)msg.Rows;
             OID = msg.OID;
         }
