+++ src/Npgsql/PostgresNotice.cs
@@ -169,6 +169,15 @@ namespace Npgsql
         /// <summary>
         /// Creates a new instance.
         /// </summary>
+        /// <remarks>
+        /// Exists for backwards compat with 4.0, has been removed for 5.0.
+        /// </remarks>
+        [Obsolete]
+        public PostgresNotice() : this(string.Empty, string.Empty, string.Empty, string.Empty) {}
+
+        /// <summary>
+        /// Creates a new instance.
+        /// </summary>
         public PostgresNotice(string severity, string invariantSeverity, string sqlState, string messageText)
         {
             Severity = severity;
